UI buttons, adding POI&tracks

+++ app/src/main/java/com/example/mt/map/layer/EditableLayer.kt
@@ -39,7 +39,7 @@ class EditableLayer(
         area: Bounds,
         rect: Rect,
         opacity: Int,
-        scale: Double
+        scale: Float
     ): Bitmap? {
         TODO("Not yet implemented")
     }
