+++ app/src/main/java/com/example/mt/map/renderer/GIEditableRenderer.kt
@@ -11,7 +11,7 @@ class GIEditableRenderer : GIRenderer() {
         area: Bounds,
         opacity: Int,
         rect: Rect,
-        scale: Double
+        scale: Float
     ): Bitmap? {
         return null
     }
